Update README.md
Fixed 3 typos
Q: Is the usage of PAC-IDs limited to liefe science / laboratories?
Q: What shall I do if the QR code representation of a PAC-ID gets to big to attach to the target item? 
A: Version numbers are not specified in order to keep PAC-IDs simple. If your version number refers to the service version, it is recommendet to append them to the category